Refactor StakeChain extensions and update StakeSceneViewModel
Moved and consolidated StakeChain extension methods from StakeChain+GemstoneSwift.swift into GemStakeChain+GemstonePrimitives.swift, removing the former file. Updated StakeSceneViewModel to use new computed properties on StakeChain for lockTime and minAmount, simplifying property access and reducing duplication.
